Goals token verification

+++ backend/controllers/goalsController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 //@desc   Get goals
 //@route  GET /api/goals
 //@access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find({});
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -19,6 +20,7 @@ const setGoals = asyncHandler(async (req, res) => {
   }
 
   const goal = await Goal.create({
+    user: req.user.id,
     text: req.body.text,
   });
 
@@ -35,6 +37,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
     return res.status(400).json({ message: "Goal not found" });
   }
 
+  const user = await User.findById(req.user.id);
+
+  //Check user
+  if (!user) {
+    return res.status(401).json({ message: "User not found" });
+  }
+
+  //Makes sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    return res.status(401).json({ message: "User not authorized" });
+  }
+
   await Goal.findByIdAndDelete(req.params.id);
   res.status(200).json({ id: req.params.id });
 });
@@ -44,9 +58,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
 //@access Private
 const updateGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
-  console.log("req.params.id", req.params.id);
+
   if (!goal) {
     return res.status(400).json({ message: "Goal not found" });
+  }
+  const user = await User.findById(req.user.id);
+
+  //Check user
+  if (!user) {
+    return res.status(401).json({ message: "User not found" });
+  }
+
+  //Makes sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    return res.status(401).json({ message: "User not authorized" });
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
